requirements.txt
asgiref==3.10.0
attrs==25.4.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
hypothesis==6.140.3
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.27.1
sortedcontainers==2.4.0
sqlparse==0.5.3
symengine==0.14.1
tzdata==2025.2
uritemplate==4.2.0